+++ src/components/UserDetailSections/ExtendedInfo/components/RequestPreferencesView/RequestPreferencesView.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import {
+  Checkbox,
   Col,
   KeyValue,
   Row,
@@ -9,26 +10,12 @@ import {
 
 import { requestPreferencesShape } from '../../../../../shapes';
 
-import styles from './RequestPreferencesView.css';
-
 class RequestPreferencesView extends Component {
   static propTypes = {
     requestPreferences: requestPreferencesShape,
     defaultServicePointName: PropTypes.string,
     defaultDeliveryAddressTypeName: PropTypes.string.isRequired,
   };
-
-  getHoldShelfStateTranslationKey() {
-    return this.props.requestPreferences.holdShelf
-      ? 'ui-users.requests.holdShelfYes'
-      : 'ui-users.requests.holdShelfNo';
-  }
-
-  getDeliveryStateTranslationKey() {
-    return this.props.requestPreferences.delivery
-      ? 'ui-users.requests.deliveryYes'
-      : 'ui-users.requests.deliveryNo';
-  }
 
   render() {
     const {
@@ -40,49 +27,36 @@ class RequestPreferencesView extends Component {
     return (
       <>
         <Row>
-          <Col xs={4}>
-            <span className={styles.heading}>
-              <FormattedMessage id="ui-users.requests.preferences" />
-            </span>
+          <Col xs={12} md={3}>
+            <KeyValue label={<FormattedMessage id="ui-users.requests.preferences" />}>
+              <>
+                <Checkbox
+                  label={<FormattedMessage id="ui-users.requests.holdShelf" />}
+                  checked={requestPreferences.holdShelf}
+                  disabled
+                />
+                <Checkbox
+                  label={<FormattedMessage id="ui-users.requests.delivery" />}
+                  checked={requestPreferences.delivery}
+                  disabled
+                />
+              </>
+            </KeyValue>
           </Col>
-        </Row>
-        <Row className={styles['request-preferences-row']}>
-          <Col xs={4} data-test-hold-shelf>
-            <FormattedMessage id={this.getHoldShelfStateTranslationKey()} />
+          <Col xs={12} md={3}>
+            <KeyValue label={<FormattedMessage id="ui-users.requests.defaultPickupServicePoint" />}>
+              {defaultServicePointName}
+            </KeyValue>
           </Col>
-          <Col xs={4} data-test-delivery>
-            <FormattedMessage id={this.getDeliveryStateTranslationKey()} />
+          <Col xs={12} md={3}>
+            <KeyValue label={<FormattedMessage id="ui-users.requests.fulfillmentPreference" />}>
+              {requestPreferences.fulfillment}
+            </KeyValue>
           </Col>
-        </Row>
-        <Row>
-          <Col xs={4}>
-            {requestPreferences.holdShelf && (
-              <KeyValue label={<FormattedMessage id="ui-users.requests.defaultPickupServicePoint" />}>
-                <span data-test-default-pickup-service-point>
-                  { defaultServicePointName || '-' }
-                </span>
-              </KeyValue>
-            )}
-          </Col>
-          <Col xs={4}>
-            {requestPreferences.delivery && (
-              <KeyValue label={<FormattedMessage id="ui-users.requests.fulfillmentPreference" />}>
-                <span data-test-fulfillment-preference>
-                  { requestPreferences.fulfillment }
-                </span>
-              </KeyValue>
-            )}
-          </Col>
-        </Row>
-        <Row>
-          <Col xsOffset={4} xs={8}>
-            {requestPreferences.delivery && (
-              <KeyValue label={<FormattedMessage id="ui-users.requests.defaultDeliveryAddress" />}>
-                <span data-test-default-delivery-address>
-                  { defaultDeliveryAddressTypeName || '-'}
-                </span>
-              </KeyValue>
-            )}
+          <Col xs={12} md={3}>
+            <KeyValue label={<FormattedMessage id="ui-users.requests.defaultDeliveryAddress" />}>
+              {defaultDeliveryAddressTypeName}
+            </KeyValue>
           </Col>
         </Row>
       </>